refactor: change entity to domain and change the imports as well

+++ src/main/java/br/com/movieflix/movieflix/service/CategoryService.java
@@ -1,15 +1,15 @@
 package br.com.movieflix.movieflix.service;
 
-import br.com.movieflix.movieflix.entity.Category;
-import br.com.movieflix.movieflix.entity.dto.category.CategoryResponse;
-import br.com.movieflix.movieflix.entity.dto.category.CategoryRequest;
-import br.com.movieflix.movieflix.entity.mapper.CategoryMapper;
+import br.com.movieflix.movieflix.domain.Category;
+import br.com.movieflix.movieflix.domain.dto.category.CategoryResponse;
+import br.com.movieflix.movieflix.domain.dto.category.CategoryRequest;
+import br.com.movieflix.movieflix.domain.mapper.CategoryMapper;
+import br.com.movieflix.movieflix.exception.notFound.ResourceNotFoundException;
 import br.com.movieflix.movieflix.repository.CategoryRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
-import java.util.Optional;
 
 @Service
 @RequiredArgsConstructor
@@ -20,8 +20,8 @@ public class CategoryService {
         return categories.stream().map(CategoryMapper::toDto).toList();
     }
 
-    public Optional<CategoryResponse> findById(Long id){
-        return repository.findById(id).map(CategoryMapper::toDto);
+    public CategoryResponse findById(Long id){
+        return repository.findById(id).map(CategoryMapper::toDto).orElseThrow(()-> new ResourceNotFoundException("Category", id));
     }
 
     public List<Category> findAllById(List<Long> categoryIds){
@@ -33,6 +33,9 @@ public class CategoryService {
     }
 
     public void deleteById(Long id){
+        if(!repository.existsById(id)){
+            throw new ResourceNotFoundException("Category", id);
+        }
         repository.deleteById(id);
     }
 }
