+++ src/main/java/br/com/movieflix/movieflix/service/MovieService.java
@@ -1,12 +1,13 @@
 package br.com.movieflix.movieflix.service;
 
-import br.com.movieflix.movieflix.entity.Category;
-import br.com.movieflix.movieflix.entity.Movie;
-import br.com.movieflix.movieflix.entity.Streaming;
-import br.com.movieflix.movieflix.entity.dto.movie.MovieRequest;
-import br.com.movieflix.movieflix.entity.dto.movie.MovieResponse;
-import br.com.movieflix.movieflix.entity.dto.movie.MovieUpdateRequest;
-import br.com.movieflix.movieflix.entity.mapper.MovieMapper;
+import br.com.movieflix.movieflix.domain.Category;
+import br.com.movieflix.movieflix.domain.Movie;
+import br.com.movieflix.movieflix.domain.Streaming;
+import br.com.movieflix.movieflix.domain.dto.movie.MovieRequest;
+import br.com.movieflix.movieflix.domain.dto.movie.MovieResponse;
+import br.com.movieflix.movieflix.domain.dto.movie.MovieUpdateRequest;
+import br.com.movieflix.movieflix.domain.mapper.MovieMapper;
+import br.com.movieflix.movieflix.exception.notFound.ResourceNotFoundException;
 import br.com.movieflix.movieflix.repository.MovieRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
@@ -28,8 +29,8 @@ public class MovieService {
         return movies.stream().map(MovieMapper::toDto).toList();
     }
 
-    public Optional<MovieResponse> findByid(Long id){
-        return repository.findById(id).map(MovieMapper::toDto);
+    public MovieResponse findById(Long id){
+        return repository.findById(id).map(MovieMapper::toDto).orElseThrow(()-> new ResourceNotFoundException("Movie", id));
     }
 
     public MovieResponse create(MovieRequest request){
@@ -42,7 +43,7 @@ public class MovieService {
 
         List<Streaming> streamings = streamingService.findAllById(request.streamings());
         if (streamings.size() != request.streamings().size()){
-            throw new RuntimeException("Some category does not exist");
+            throw new RuntimeException("Some streaming does not exist");
         }
          Movie movie = Movie.builder()
                 .name(request.name())
@@ -57,7 +58,7 @@ public class MovieService {
     }
 
     public MovieResponse update(Long id, MovieUpdateRequest request){
-        Movie foundMovie = repository.findById(id).orElseThrow(()->new RuntimeException("Movie not found"));
+        Movie foundMovie = repository.findById(id).orElseThrow(()-> new ResourceNotFoundException("Movie", id));
         if(request.name() != null )foundMovie.setName(request.name());
         if(request.description() != null )foundMovie.setName(request.name());
         if(request.releaseDate() != null )foundMovie.setName(request.name());
@@ -81,15 +82,25 @@ public class MovieService {
 
     }
     public void deleteById(Long id){
+
+        if(!repository.existsById(id)){
+            throw new ResourceNotFoundException("Movie", id);
+        }
         repository.deleteById(id);
     }
 
     public List<MovieResponse> findMovieByCategoryId(Long categoryId){
+        if(!repository.existsById(categoryId)){
+            throw new ResourceNotFoundException("Movie", categoryId);
+        }
         List<Movie> moviesByCategoryIds = repository.findMovieByCategories_Id(categoryId);
         return moviesByCategoryIds.stream().map(MovieMapper::toDto).toList();
     }
 
     public List<MovieResponse> findMovieByStreamingId(Long streamingId){
+        if(!repository.existsById(streamingId)){
+            throw new ResourceNotFoundException("Streaming", streamingId);
+        }
         List<Movie> moviesByStreamingId = repository.findMovieByStreamings_Id(streamingId);
         return moviesByStreamingId.stream().map(MovieMapper::toDto).toList();
     }
